test: enable CSSInlineEdit-integ-test.js
